requirements.txt
fastapi 
uvicorn 
requests 
PyPDF2 
python-docx 
sentence-transformers 
faiss-cpu 
groq 
supabase